+++ src/protocol/mod.rs
@@ -21,7 +21,8 @@ use crate::task::{Signature, Task};
 static ORIGIN: Lazy<Option<String>> = Lazy::new(|| {
     hostname::get()
         .ok()
-        .map(|sys_hostname| format!("gen{}@{:?}", process::id(), sys_hostname))
+        .and_then(|sys_hostname| sys_hostname.into_string().ok())
+        .map(|sys_hostname| format!("gen{}@{}", process::id(), sys_hostname))
 });
 
 /// Create a message with a custom configuration.
@@ -266,7 +267,7 @@ pub struct MessageHeaders {
     /// A string representation of the keyword arguments of the task.
     pub kwargsrepr: Option<String>,
 
-    /// A string representing the node that produced the task.
+    /// A string representing the nodename of the process that produced the task.
     pub origin: Option<String>,
 }