+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+- `Message::headers::origin` field fixed. Before it included quotes around the hostname.
+
+### Added
+
+- `Request::hostname` field now populated by the `Celery` app consuming the task.
+
 ## [0.2.5] - 2019-03-24
 
 ### Changed